+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - TMA-512 - Fix invalid data type issue in blueprint when a field has > 255 characters
 - TMA-506 - Add option :time to report.execute to force platform simulate the result of report at that time
 - TMA-509 - Provisioning: Should clean all zombie clients when project creation limit reached
+- TMA-542 - Release: Should fail when segments contains duplicated ids
 
 ## 0.6.53
 - TMA-522 - Rollout: Incorrect CLIENT_ID assigned to client schedule